+++ src/game/stages/play/components/animated.hpp
@@ -8,7 +8,7 @@ struct Animated
     Animated(const sf::Vector2i& frameSize, const int framesNumber, const float speed)
         : frames{}, frame{}, speed{speed}, paused{false}
     {
-        frames.reserve(framesNumber);
+        frames.reserve(static_cast<std::size_t>(framesNumber));
 
         for (int i = 0; i < framesNumber; i++)
             frames.emplace_back(sf::IntRect(i * frameSize.x, 0, frameSize.x, frameSize.y));
